refactor getting database components -> ClusterNode
