+++ src/data/ClusterNode.js
@@ -1,4 +1,5 @@
 import Parser from 'uri-parser';
+import * as Sentry from '@sentry/browser';
 
 /**
  * Abstraction that captures details and information about a node in a cluster.
@@ -12,6 +13,19 @@ export default class ClusterNode {
         this.addresses = record.get('addresses');
         this.role = record.get('role');
         this.database = record.get('database');
+        this.dbms = {};
+    }
+
+    asJSON() {
+        return {
+            address: this.getBoltAddress(),
+            procotols: this.protocols(),
+            role: this.role,
+            database: this.database,
+            id: this.id,
+            label: this.getLabel(),
+            dbms: this.dbms,
+        };
     }
 
     getBoltAddress() {
@@ -42,5 +56,69 @@ export default class ClusterNode {
         return this.addresses
             .map(addr => Parser.parse(addr))
             .map(parsed => parsed.protocol);
+    }
+
+    /**
+     * Returns true if the context is attached to a Neo4j Enterprise edition server,
+     * false otherwise.
+     */
+    isEnterprise() {
+        return this.dbms.edition === 'enterprise';
+    }
+
+    /**
+     * Returns true if the context provides for native auth management, false otherwise.
+     */
+    supportsNativeAuth() {
+        return this.dbms.nativeAuth;
+    }
+
+    checkComponents(driver) {
+        const q = 'call dbms.components()';
+        const session = driver.session();
+
+        const componentsPromise = session.run(q, {})
+            .then(results => {
+                const rec = results.records[0];
+                this.dbms.name = rec.get('name')
+                this.dbms.versions = rec.get('versions');
+                this.dbms.edition = rec.get('edition');
+            })
+            .catch(err => {
+                Sentry.captureException(err);
+                console.error('Failed to get DBMS components');
+                this.dbms.name = 'UNKNOWN';
+                this.dbms.versions = [];
+                this.dbms.edition = 'UNKNOWN';
+            });
+
+        // See issue #27 for what's going on here.  DB must support native auth
+        // in order for us to expose some features, such as user management.
+        const authQ = `
+            CALL dbms.listConfig() YIELD name, value 
+            WHERE name =~ 'dbms.security.auth_provider.*' 
+            RETURN value;`;
+        const authPromise = session.run(authQ, {})
+            .then(results => {
+                let nativeAuth = false;
+
+                results.records.forEach(rec => {
+                    const val = rec.get('value');
+                    const valAsStr = `${val}`; // Coerce ['foo','bar']=>'foo,bar' if present
+
+                    if (valAsStr.indexOf('native') > -1) {
+                        nativeAuth = true;
+                    }
+                });
+
+                this.dbms.nativeAuth = nativeAuth;
+            })
+            .catch(err => {
+                Sentry.captureException(err);
+                console.error('Failed to get DBMS auth implementation type');
+                this.dbms.nativeAuth = false;
+            });
+
+        return Promise.all([componentsPromise, authPromise]);
     }
 }
